+++ lib/messagebox.c
@@ -77,7 +77,7 @@ message_autosize(struct bsddialog_conf *conf, int rows, int cols, int *h, int *w
 	if (rows == BSDDIALOG_AUTOSIZE) {
 		*h = MIN_HEIGHT - 1;
 		if (maxword > 0)
-			*h += MAX(nlines, (*w / GET_ASPECT_RATIO(conf)));
+			*h += MAX(nlines, (int)(*w / GET_ASPECT_RATIO(conf)));
 		*h = MAX(*h, MIN_HEIGHT);
 		/* avoid terminal overflow */
 		*h = MIN(*h, widget_max_height(conf));
